chore: remove core folder from source hierarchy
